Extract out test to abstract class (so it can be used for linux/android
